+++ designsystem/src/main/java/com/mundcode/designsystem/theme/Theme.kt
@@ -6,26 +6,12 @@ import androidx.compose.material.darkColors
 import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
 
-private val DarkColorPalette = darkColors(
+val DarkColorPalette = darkColors(
     background = Gray900,
     onBackground = White
 )
 
-private val LightColorPalette = lightColors(
+val LightColorPalette = lightColors(
     background = White,
     onBackground = Gray900
 )
-
-@Composable
-fun MuntamTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Composable () -> Unit) {
-    val colors = if (darkTheme) {
-        DarkColorPalette
-    } else {
-        LightColorPalette
-    }
-
-    MaterialTheme(
-        colors = colors,
-        content = content
-    )
-}
